requirements.txt
dj-database-url==0.5.0
Django==1.11.6
django-allauth==0.36.0
django-rest-knox==3.4.0
djangorestframework==3.9.0
gunicorn==19.9.0
matplotlib==3.0.2
psycopg2==2.7.3.2
python-dateutil==2.7.5
python3-openid==3.1.0
requests==2.19.1
urllib3==1.23
vine==1.1.4
whitenoise==4.0
xlrd==1.1.0
